Update classes and inheritance
Added a new property to floors, declared in the constructor area (where "residents" is declared), don't need to declare this in the parent class because round huts don't have floors. therefore, this is unique to the current class. 

The system should automatically run a multiplication and the updated capacity should be the product of floors and whatever there was before 

🤔 when I was attempting it I thought it was a val to be declared but NO! because this is a parameter (understood as a characteristic of this class?) so it is added in the constructor section.